+++ Assets/Scripts/Games/DrummingPractice/DrummingPractice.cs
@@ -16,7 +16,7 @@ namespace HeavenStudio.Games.Loaders
             {
                 new GameAction("bop", "Bop")
                 {
-                    function = delegate { var e = eventCaller.currentEntity; DrummingPractice.instance.SetBop((float) e.beat, e.length, e["bop"], e["autoBop"]); }, 
+                    function = delegate { var e = eventCaller.currentEntity; DrummingPractice.instance.SetBop(e.beat, e.length, e["bop"], e["autoBop"]); }, 
                     resizable = true,
                     parameters = new List<Param>()
                     {
@@ -26,7 +26,7 @@ namespace HeavenStudio.Games.Loaders
                 },
                 new GameAction("drum", "Hit Drum")
                 {
-                    function = delegate { var e = eventCaller.currentEntity; DrummingPractice.instance.Prepare((float) e.beat, e["toggle"]); }, 
+                    function = delegate { var e = eventCaller.currentEntity; DrummingPractice.instance.Prepare(e.beat, e["toggle"]); }, 
                     defaultLength = 2f, 
                     parameters = new List<Param>()
                     {
@@ -47,7 +47,7 @@ namespace HeavenStudio.Games.Loaders
                 },
                 new GameAction("move npc drummers", "NPC Drummers Enter or Exit")
                 {
-                    function = delegate {var e = eventCaller.currentEntity; DrummingPractice.instance.NPCDrummersEnterOrExit((float) e.beat, e.length, e["exit"], e["ease"]); },
+                    function = delegate {var e = eventCaller.currentEntity; DrummingPractice.instance.NPCDrummersEnterOrExit(e.beat, e.length, e["exit"], e["ease"]); },
                     defaultLength = 4f,
                     resizable = true,
                     parameters = new List<Param>()
@@ -109,7 +109,7 @@ namespace HeavenStudio.Games
 
         [Header("Variables")]
         float movingLength;
-        float movingStartBeat;
+        double movingStartBeat;
         bool isMoving;
         string moveAnim;
         EasingFunction.Ease lastEase;
@@ -164,7 +164,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void NPCDrummersEnterOrExit(float beat, float length, bool exit, int ease)
+        public void NPCDrummersEnterOrExit(double beat, float length, bool exit, int ease)
         {
             movingStartBeat = beat;
             movingLength = length;
@@ -177,7 +177,7 @@ namespace HeavenStudio.Games
             });
         }
 
-        public void SetBop(float beat, float length, bool shouldBop, bool autoBop)
+        public void SetBop(double beat, float length, bool shouldBop, bool autoBop)
         {
             goBop = autoBop;
             if (shouldBop)
@@ -199,7 +199,7 @@ namespace HeavenStudio.Games
             rightDrummer.Bop();
         }
 
-        public void Prepare(float beat, bool applause)
+        public void Prepare(double beat, bool applause)
         {
             int type = count % 2;
             player.Prepare(beat, type);
